Gui build V2
Added the page viewer with 4 images in an array sliding with swipes, added on the onlick listener for the login button. Also added an instagram icon for testing that re-directs users to the instagram url via on click. Changed up the launcher icon and background color. Fixed constraints on customer registration activity.